Extracted utils from `utils/changelog` to separate files.
