+++ packages/ckeditor5-dev-release-tools/lib/utils/constants.js
@@ -1,0 +1,14 @@
+/**
+ * @license Copyright (c) 2003-2024, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md.
+ */
+
+/**
+ * Changelog file name.
+ */
+export const CHANGELOG_FILE = 'CHANGELOG.md';
+
+/**
+ * Changelog header.
+ */
+export const CHANGELOG_HEADER = 'Changelog\n=========\n\n';
