+++ packages/ckeditor5-dev-release-tools/tests/tasks/creategithubrelease.js
@@ -5,7 +5,7 @@
 
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import createGithubRelease from '../../lib/tasks/creategithubrelease.js';
-import transformCommitUtils from '../../lib/utils/transformcommitutils.js';
+import * as transformCommitUtils from '../../lib/utils/transformcommitutils.js';
 
 const stubs = vi.hoisted( () => ( {
 	constructor: vi.fn(),
@@ -26,7 +26,7 @@ vi.mock( '@octokit/rest', () => ( {
 	}
 } ) );
 
-vi.mock( '../../lib/utils/transformcommitutils' );
+vi.mock( '../../lib/utils/transformcommitutils.js' );
 
 describe( 'createGithubRelease()', () => {
 	let options;
